+++ alibi_detect/cd/mmd.py
@@ -2,7 +2,7 @@ import dask.array as da
 from functools import partial
 import logging
 import numpy as np
-from typing import Callable, Dict, Tuple, Union
+from typing import Callable, Dict, Optional, Tuple, Union
 from alibi_detect.base import BaseDetector, concept_drift_dict
 from alibi_detect.cd.preprocess import preprocess_drift
 from alibi_detect.cd.utils import update_reference
@@ -19,14 +19,15 @@ class MMDDrift(BaseDetector):
                  p_val: float = .05,
                  X_ref: Union[np.ndarray, list] = None,
                  preprocess_X_ref: bool = True,
-                 update_X_ref: Dict[str, int] = None,
-                 preprocess_fn: Callable = None,
-                 preprocess_kwargs: dict = dict(),
+                 update_X_ref: Optional[Dict[str, int]] = None,
+                 preprocess_fn: Optional[Callable] = None,
+                 preprocess_kwargs: Optional[dict] = None,
                  kernel: Callable = gaussian_kernel,
-                 kernel_kwargs: dict = None,
+                 kernel_kwargs: Optional[dict] = None,
                  n_permutations: int = 1000,
-                 chunk_size: int = None,
-                 data_type: str = None
+                 chunk_size: Optional[int] = None,
+                 input_shape: Optional[tuple] = None,
+                 data_type: Optional[str] = None
                  ) -> None:
         """
         Maximum Mean Discrepancy (MMD) data drift detector using a permutation test.
@@ -56,6 +57,8 @@ class MMDDrift(BaseDetector):
             Number of permutations used in the permutation test.
         chunk_size
             Chunk size if dask is used to parallelise the computation.
+        input_shape
+            Shape of input data.
         data_type
             Optionally specify the data type (tabular, image or time-series). Added to metadata.
         """
@@ -64,16 +67,16 @@ class MMDDrift(BaseDetector):
         if p_val is None:
             logger.warning('No p-value set for the drift threshold. Need to set it to detect data drift.')
 
-        if isinstance(preprocess_kwargs, dict) and not isinstance(preprocess_fn, Callable):  # type: ignore
-            preprocess_fn = preprocess_drift
-
-        if isinstance(preprocess_fn, Callable):  # type: ignore
+        if isinstance(preprocess_kwargs, dict):  # type: ignore
+            if not isinstance(preprocess_fn, Callable):  # type: ignore
+                preprocess_fn = preprocess_drift
             self.preprocess_fn = partial(
                 preprocess_fn,
                 **preprocess_kwargs
             )
+            keys = list(preprocess_kwargs.keys())
         else:
-            self.preprocess_fn = None
+            self.preprocess_fn, keys = None, []
 
         # optionally already preprocess reference data
         self.preprocess_X_ref = preprocess_X_ref
@@ -82,6 +85,13 @@ class MMDDrift(BaseDetector):
         self.n = X_ref.shape[0]  # type: ignore
         self.p_val = p_val
         self.chunk_size = chunk_size
+
+        if isinstance(input_shape, tuple):
+            self.input_shape = input_shape
+        elif 'max_len' in keys:
+            self.input_shape = (preprocess_kwargs['max_len'],)
+        elif isinstance(X_ref, np.ndarray):
+            self.input_shape = X_ref.shape[1:]
 
         kwargs = kernel_kwargs if isinstance(kernel_kwargs, dict) else {}
         kwargs['kernel'] = kernel